LayoutTests/inspector/debugger/setShouldBlackboxURL-source-map-inner-expected.txt

Tests Debugger.setShouldBlackboxURL.


== Running test suite: Debugger.setShouldBlackboxURL
-- Running test case: Debugger.setShouldBlackboxURL.SourceMap.Pause.inner.Blackbox.inner
Fetching scripts...
Blackboxing script...
Finding pause location...
Creating breakpoint...
Adding breakpoint...
Triggering pause...
PAUSE AT n:3:8
      0    import { inner } from "./inner.js"
      1
 ->   2    export |function middle(x) {
      3        let y = "middle";
      4        y += inner(x);
      5        return y;

PAUSE AT t:6:12
      2    export function middle(x) {
      3        let y = "middle";
      4        y += inner(x);
 ->   5        return |y;
      6    }
      7

PAUSE AT t:3:1
      0    import { middle } from "./middle.js"
      1
 ->   2    |globalThis.outer = function(x) {
      3        let y = "outer";
      4        y += middle(x);
      5        return y;

PAUSE AT <anonymous>:6:12
      2    globalThis.outer = function(x) {
      3        let y = "outer";
      4        y += middle(x);
 ->   5        return |y;
      6    };
      7

PAUSE AT <anonymous>:7:1
      3        let y = "outer";
      4        y += middle(x);
      5        return y;
 ->   6    |};
      7

PAUSE AT Global Code:1:10
--- Source Unavailable ---

Removing breakpoint...
Unblackboxing script...

-- Running test case: Debugger.setShouldBlackboxURL.SourceMap.Pause.inner.Blackbox.middle
Fetching scripts...
Blackboxing script...
Finding pause location...
Creating breakpoint...
Adding breakpoint...
Triggering pause...
PAUSE AT n:2:5
      0    export function inner(x) {
 ->   1        |return "inner" + x;
      2    }
      3

PAUSE AT t:3:1
      0    import { middle } from "./middle.js"
      1
 ->   2    |globalThis.outer = function(x) {
      3        let y = "outer";
      4        y += middle(x);
      5        return y;

PAUSE AT <anonymous>:6:12
      2    globalThis.outer = function(x) {
      3        let y = "outer";
      4        y += middle(x);
 ->   5        return |y;
      6    };
      7

PAUSE AT <anonymous>:7:1
      3        let y = "outer";
      4        y += middle(x);
      5        return y;
 ->   6    |};
      7

PAUSE AT Global Code:1:10
--- Source Unavailable ---

Removing breakpoint...
Unblackboxing script...

-- Running test case: Debugger.setShouldBlackboxURL.SourceMap.Pause.inner.Blackbox.outer
Fetching scripts...
Blackboxing script...
Finding pause location...
Creating breakpoint...
Adding breakpoint...
Triggering pause...
PAUSE AT n:2:5
      0    export function inner(x) {
 ->   1        |return "inner" + x;
      2    }
      3

PAUSE AT n:3:8
      0    import { inner } from "./inner.js"
      1
 ->   2    export |function middle(x) {
      3        let y = "middle";
      4        y += inner(x);
      5        return y;

PAUSE AT t:6:12
      2    export function middle(x) {
      3        let y = "middle";
      4        y += inner(x);
 ->   5        return |y;
      6    }
      7

PAUSE AT Global Code:1:10
--- Source Unavailable ---

Removing breakpoint...
Unblackboxing script...
